requirements.txt
absl-py==1.1.0
albumentations==1.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
easydict==1.9
entrypoints==0.4
executing==0.8.3
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.2
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
networkx==2.8.4
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.1.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.2.0
Shapely==1.8.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.17.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
trio==0.20.0
trio-websocket==0.9.2
typeguard==2.13.3
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webdriver-manager==3.7.0
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
